Another BUG FOUND, because the genAtlas method attempts to add all textures ending with.png into a new atlas texture, the old atlas texture is merged with the current one…. I'm not sure how this works.

+++ ToolScripts/InGamePythonInterpreter4.cs
@@ -431,7 +431,7 @@ public void iterateColliderSides (GameObject wall)
 			
 		}
 			DestroyImmediate(tempParent);
-			vertexCentroid = (vertexCentroid/currentVerts.Count);
+			vertexCentroid = (vertexCentroid/(float)currentVerts.Count);
 			return vertexCentroid;
 	}
